# HutChil-POS/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HutChil POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List employees
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: open
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: createShift
      tags: [shifts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}:
    get:
      operationId: getShift
      tags: [shifts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}/close:
    post:
      operationId: closeShift
      tags: [shifts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftCloseInput"
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /brewing:
    post:
      operationId: logBrewing
      tags: [shifts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrewingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrewingLog"
  /stock/logs:
    get:
      operationId: listStockLogs
      tags: [stock]
      security:
        - bearerAuth: []
      parameters:
        - name: product_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockLog"
...
  /sheets/settings:
    get:
      operationId: getSheetsSettings
      tags: [sheets]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetsSettings"
    put:
      operationId: updateSheetsSettings
      tags: [sheets]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SheetsSettingsInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetsSettings"
  /sheets/export:
    post:
      operationId: exportToSheets
      tags: [sheets]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetsExportResult"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, name, username, role]
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          type: string
        daily_wage:
          type: ["number", "null"]
        shift:
          type: ["string", "null"]
        created_at:
          type: string
    UserInput:
      type: object
      required: [name, username, password, role]
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          type: string
        daily_wage:
          type: ["number", "null"]
        shift:
          type: ["string", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: ["string", "null"]
        role:
          type: string
        daily_wage:
          type: ["number", "null"]
        shift:
          type: ["string", "null"]
...
    Shift:
      type: object
      required: [id, user_id, shift_type, check_in_time, status]
      properties:
        id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: ["string", "null"]
        shift_type:
          type: string
        check_in_time:
          type: string
        close_time:
          type: ["string", "null"]
        status:
          type: string
        cash_count:
          type: ["number", "null"]
        account_total:
          type: ["number", "null"]
        system_cash:
          type: ["number", "null"]
        system_transfer:
          type: ["number", "null"]
        system_total:
          type: ["number", "null"]
        net_total:
          type: ["number", "null"]
        difference:
          type: ["number", "null"]
        photos:
          type: ["string", "null"]
        brewing_product_id:
          type: ["integer", "null"]
        brewing_qty:
          type: ["integer", "null"]
    ShiftInput:
      type: object
      required: [shift_type]
      properties:
        shift_type:
          type: string
        brewing_product_id:
          type: ["integer", "null"]
        brewing_qty:
          type: ["integer", "null"]
    ShiftCloseInput:
      type: object
      required: [cash_count, account_total, photos]
      properties:
        cash_count:
          type: number
        account_total:
          type: number
        photos:
          type: array
          items:
            type: string
    BrewingInput:
      type: object
      required: [shift_id, product_id, qty]
      properties:
        shift_id:
          type: integer
        product_id:
          type: integer
        qty:
          type: integer
    BrewingLog:
      type: object
      required: [id, shift_id, product_id, qty, time]
      properties:
        id:
          type: integer
        shift_id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: ["string", "null"]
        qty:
          type: integer
        time:
          type: string
    StockLog:
      type: object
      required: [id, product_id, qty_change, reason, type, time]
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: ["string", "null"]
        qty_change:
          type: integer
        reason:
          type: string
        type:
          type: string
        user_id:
          type: ["integer", "null"]
        user_name:
          type: ["string", "null"]
        time:
          type: string
...
    SheetsSettings:
      type: object
      required: [id]
      properties:
        id:
          type: integer
        spreadsheet_id:
          type: ["string", "null"]
        service_account_json:
          type: ["string", "null"]
        auto_sync:
          type: boolean
        last_sync:
          type: ["string", "null"]
    SheetsSettingsInput:
      type: object
      properties:
        spreadsheet_id:
          type: ["string", "null"]
        service_account_json:
          type: ["string", "null"]
        auto_sync:
          type: boolean
    SheetsExportResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        last_sync:
          type: ["string", "null"]