# HutChil-POS/lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HutChil POS API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  username?: string;
  /** @nullable */
  password?: string | null;
  role?: string;
  /** @nullable */
  daily_wage?: number | null;
  /** @nullable */
  shift?: string | null;
}